GenAi/Day1/requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-openai
langchain-groq
langchain-mistralai
langchain-google-genai
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests
langchain_huggingface
transformers